Local CO-OP edition Part 2
A lot of things need to be fixed though. Like a lot

+++ MyGame/SpaceGame/Position2D.cpp
@@ -4,7 +4,7 @@
 Position2D::Position2D()
 {
 }
-Position2D::Position2D(int x, int y) :_x(x), _y(y)
+Position2D::Position2D(float x, float y) :_x(x), _y(y)
 {
 }
 
@@ -13,9 +13,9 @@ Position2D::~Position2D()
 {
 }
 
-int Position2D::getX() { return _x; }
-int Position2D::getY() { return _y; }
+float Position2D::getX() { return _x; }
+float Position2D::getY() { return _y; }
 
 
-void Position2D::setX(int x) { _x = x; }
-void Position2D::setY(int y) { _y = y; }
+void Position2D::setX(float x) { _x = x; }
+void Position2D::setY(float y) { _y = y; }
